Feature - implement ```{foldable} Title ...``` syntax.
Implemented wiki markup language syntax for

````markdown
```{foldable}  Title of foldable element

```
````

Note that this code block does not include the (````) four backticks.

That code block is rendered to the following html code

````html
<details>
  <summary>
     <u class="solution-label">Example</u>
  <summary>
       ... ... ... ...
     ... ... .. ...
</details>
````

This syntax is similar to ```{solution} ... ``` and ```{proof} ....``` special code blocks, but it allows changing the label of <summary> html block.