Feature - implement wordlinks for frontmatter.
The wordlinks feature allows quickly creating hyperlinks for words
 defined in the wordlinks section of the frontmatter, YAML metadata
within (---) at the beginning of a wiki page. It is useful for creating
hyperlinks to dictionry definitions or Wikipedia articles related to
a particular word.  For instance consider the following frontmatter
of a Wiki page *.md file.

```
---
description: Wiki page (note) description here.
abbreviations:
    HR:  Human Resources
    GUI: Graphics User Interface
wordlinks:
   ZFS:              https://hyperlink-to-wikipedia-of-ZFS
   TrueNAS:   https://link-to-wikipedia-of-TrueNAS
---
```

Whenever the word ZFS or TrueNAS appears in any MWiki,
paragraph, quote or bullet list, it is replaced by a hyperlink whose
label is the word and the URL is the corresponding URL defined
wordlist YAML dictionary. For instance, if a text has the  word ZFS
as the text

```md
The file system ZFS is awesome!
```

which is rendered to HTML as

```html
The file system
<a href=" https://hyperlink-to-wikipedia-of-ZFS" ...>ZFS</a>
is awesome.
```