Feature - implement /tags page.
Implement /tags page, which lists all hashtags, and a tag
extraction background service (subprocess) . This feature
increases the document discoverability. For instance, if a the wiki
contains the hashtags #authoring, #writing and #essary, the /tags
page will contain hyperlinks with those tags which search them
when the user clickts at any hashtag hyperlink.