Feature implement file upload.
This commit implements the button with label 'Link to Uploaded File' for
the wiki markdown code/text editor. When this button is clicked, a upload popup
window is shown, where the user can select a file to be uploaded to the wiki server
folder <WIKIPATH>/upload. Upon successful upload, a hyperlink to the file is inserted
at the cursor position in the code editor.