Feature - Implement document preview button.
The feature introduced by this commit provides a document preview
button in the wiki code editor, which allows the user to  view a preview of the current wiki page being edited without saving it. The preview button
is a common Wiki feature found in Media Wiki, DocuWiki and other engines,
that was missing in MWiki.