Feature - implement adjustable LaTeX equation enumeration.
Implement 'equation_enumeration' directive for a front-matter of a MWiki page
(markdown file). This directive allows enabling/disabling LaTeX equation enumeration
and changing equation enumeration style. If the directive is set to 'none', automatic
equation enumeration is disabled. If the directive is set to 'section' the enumeration
 schema becomes `<section-number>.<equation-number>` (default style). If the directive
 is set to to 'subsection' , the enumeration style is set to  `<section>.<subsection>.<eqnum>`.

For instance, if a LaTeX (MathJaX) equation is numbered as 2.4.10, it means that is the 10th
equation of the 4th subsection of the section 2 of the current wiki page.