Feature - allow changing document title in wiki page frontmatter.
By default, the wiki page title is the corresponding markdown file name.
This feature allows changing the title in the wiki page frontmatter.
For instance if the document name is 'Computer Simulation.org'
and the frontmatter is

```markdown
---
title:  Simulation of Mechanical Systems - Solving ODEs
---
... ... ... ... ..

## Section 1 ....
 ... ... ...
````

The document title will be changed to "Simulation of Mechanical Systems - Solving ODEs".