Feature - implement video lazy loading.
The feature introduced by this commit allows faster
page loading due to the delayed loading of
embedded mp4 and webm videos, which are only
loaded when visible in the current window (viewport). Other benefit is the reduced load
on the server as pages with many videos does
not cause many immediate http requests.