Bugfix - Fix preview of images/figures and video.
Issue: Due to code changes, the editor preview button, which
displays how the code of wiki page will look like before saving,
was not showing embedded pictures/images or videos. The bug
was introduced after the lazy loading feature.

Solution: Refactor class HtmlRederer and add a flag preview (boolean)
member variable, that when set to true, the page is rendered in preview
mode in which image, figures and videos are not rendered with lazy loading.