Wiki pages (notes) with spaces in the file name are regarded as the same as pages with underline in the file name.
After this change, the files  'Cloud Computing.md' abd 'Cloud_Computing.md'
are regarded as the same file, consequentely the wiki pages 'Cloud Computing'
and 'Cloud_Computing' are also regarded as the wiki page (note). This approach
improves compatibility with Wikis or note taking applications  that use underline
in file name and Wikis that use space in the file name.