Bugfix - Fix the display of edit buttons (pencil icons) due to architecture changes.
Passing server data to the page via template no longer works
because cached pages are served directly from the compilation
output (html files). As a result, a rest api endpoint /api/auth was
implemented for allowing the client-side to obtain the configuration
data for modify the page without markdown recompilation on every
request.