Feature - Implement show licenses checkbox in settings form.
Implment checkbox show licenses (enabled by default) in the settings form.
If this checkbox is eanbled, a menu option whith the label 'Licenses' is added to
the main menu. When clicked this menu option redirects to a page showin the
license of this project AGPL (GNU Afero) and the licenses of all depedencies
used by this project giving credit to other projects and open source developers.