Recompile markdown file if any link depency is changed.
Details: If a page contains an internal to hyperlink to a Wiki page
[Mass Moment of Internetia Tensor]] and the corresponding
markdown page does not exist yet the link will be red.
If this page corresponding to this link is created and the
page referencing this link is refreshed, it will be recompiled for
updating the cached html. After recompilation the link will
become green.