Feature implement custom MyST role for adding notes to text.
Implement MyST {note}`[TERMS OR SENTENCE] ([NOTE])` that is
displayed as the text [TERMS or SENTENCE] with underline dots.
When the user clicks at this text a popup window displaying the
[NOTE] containing notes or observation is shown to the user.
For instance, the code

````markdown
A python {note}`dictionary (This data structure is a hash table)` has
type Dict[key, value] and can be created using curly braces.
````

is displayed as: "A python dictionary ...."  where the term dictionary
is shown with underline dots. When the user clicks at this term,
a popup window with the title "Note" and content "This data structure ..."
is shown to the user.