Extend syntax of foldable blocks {example}, {solution} and {proof}.
The parser was extended to support the foldable block syntax

````
```{solution} Solution using Lagragian Mechanics approach
Describe the solution here.
```
````

Now this block is compiled to html as

```html
<div>
   <details>
      <summary>
          <label data-i18n="foldable-math-solution-block-label">Solution</label>
          Solution using Lagragian Mechanics approach
      </summary>
        <!-- Content here-->
   </details>
</div>
```

The text of the label element is localized to the user language
provided by the web browser or the default system language.