Improve search - normalize Vietnamese latin alphabet.
Normalize letters, symbols and diactrics of the Vietnamese latin alphabet to English alphabet for improving the search engine.
This feature allows searching for vietnamese words using
an English keyboard layout. For instance, the vietnamese word
"người" is stored in the search index database as "nguoi" instead
of being stored as "người". So, searching for "nguoi" or "người"
yields the same search results. As a result, this approach makes
it easier to search for vietnamese words using an English keyboard
layout or non vietnamese keyboard layouts.