Improve search - normalize escape characters \. and \:.
The domain name of websites such as somedomain.org is parsed
as URL by default. When this behavior is not desirable, the URL is
written as somedomain\.org. However, this breaks the search as
searching for somedomain.org yields no result. The solution for
this issue added by this commit is to to turn the escaped character
"\." into "." and "\:" into ":" when storing the text in the search index
database for full text search.