Feature - Implement rendering of MyST markdown footnotes.
This commit implements rendering of the following  MyST markdown
syntax for footnotes to html.

````markdown
Some text here term^[footnote 1 explanation here] and continuation
second term^[here is the content of footnote 2] end of paragraph.
````

It is rendered to

````markdown
                                     [1]                                                        [2]
Some text here term       and continuation second term
end of paragraph.
````

Where [1] and [2] are superspcript hyperlinks in square brackets
that displays the footnote content in a popup window when
clicked.