Feature - Extend syntax of {comment} block for toggling rendering of enclosed code.
The {comment} syntax was extended with commands "on" and "off"
for toggling the rendering/display of the enclosed markdown code
without needing the removal of ```{comment}...``` block. This feature is
useful for toggling the display of content that is not ready to be published.

Example 1: The enclosed code is not rendered/displayed
and regarded a comment.

````markdown
```{comment} off
Enclosed code here and [some hyperlink](https://somepage.com).
```
```

Example 2: The enclosed code is displayed after the page is saved.

````markdown
```{comment} on
Enclosed code here and [some hyperlink](https://somepage.com).
```
````