Feature - Implement MyST role raw {r}`text` and {raw}`text`.
Due to compatibility with Github Flavored Markdown, angle brackets
may be interpreted as html tags which may not be desirable. MyST raw role
{r}`enclosed text` or {raw}`enclosed text` solves this issue by rendering the
text to iself. For instance the markdown {r}`<a href="https://site.com">label</a>`
code is rendered to <a href="https://site.com">label</a> without monospace
font/typeafce instead of being rendered to a html hyperlink.