Refactor footnotes parser and renderer - replace footnotes parser with custom one.
The markdown_it_py footnotes parser presented a buggy behavior
that could not be fixed. As a result, it was replaced with a custom
markdown_it plugin for parsing footnotes. The footnotes syntax also
was changed to `^{footnote syntax content}`, for instance

````markdown
The motor speed is about 3000 rpm^{revolutions per minute}.
````

The syntax to foonotes references was changed because
the previous one `^[foonote content here]` conflicted with
existing constructs such as `[[internal link]]` or [label](http://url.to.site)`.