Secret key is now stored in the wiki repository.
Before this commit, the secret key was stored in special folders
for application data sepecific for each operating system. Now
the secret key is stored in the file $WIKI_REPOSITORY/.data/appkey.txt.
The key is a uniquely generated random hexadecimal string during the
wiki initialization and is unique per MWiki deployment.