Feature  - implement magic link for passwordless authentication.
The command $ mwiki auth generates a sign in link that also allows
authentication without password by copying and pasting the link.
The magic login link is valid for 1 minute and uses HMAC signature
to ensure its authenticity and integrity.  This feature is similar to Jupyter
server token URL login, that allows logging in to a Jupyter Lab server
by pasting the magic link in some web browser.