Improve token login process by using the token as a base64 string.
Change the token login process to use the authentication token as a base64
encoded string. Now the token is the following a dictionary encoded to JSON and base64.

````javascript
{
    // User to be authenticated
     "user": "UserAccountName"
    // Unix Timestamp UCT NOW + 1 Minute
   , "expiration:  1000022
    // Random integer
   , "salt":  20815
    // HMAC signature of the string "<USER>/<expiration>/<salt> using
    // the app private key
   , "signature": "Hmac signature"
}
````

The authencation token can be generated using the command

```sh
$ mwiki auth --user=<USERNAME> --wikipath=</path/to/markdownfiles>
```

The advantage of the authentication token is the ability to log in
without password in similar way to Jupyter lab or Jupyter notebook.