Feature - implement popup window that allows users to override the current locale.
This commit improves the i18n internationalizatioin infrastructure by  addding
the item "Language" to the main menu for opening a form that allows users
to quickly switch the current user interface language. The user language setting
is unique for each user and persisted in the browser local storage. If there is
no user locale/language setting in the local storage, then the system attempts to
get the user locale from the JavaScript navigator.language API when the checkbox
"always use default language/locale" is disabled. When the previous checkbox is enabled,
the UI language is the default one set by the admin in the settings form. Note that
even if the checkbox "always use the default language" is enabled, each user is still
able to switch the user interface language during his or her session.