Feature - add flag --auth to the command $ mwiki server command for passwordless login.
When the flag --auth is added to the command $ mwiki server, a login token and login
URL for passwordless authentication are printed in ther terminal before starting the server.
Both the token and the URL are valid for 20 seconds and can be used for fast authentication
as admin without password. The authentication is endpoint cannot be manipulated be non
authorized users as it uses a random numer salt and HMAC digital signature.