Feature - add flag --embed-mathjax to the command $ mwiki compile.
Add command line switch --embed-mathjax to the static
website generator subcommand $ mwiki compile for
self hosting mathjax in the static websiste instead of
loading the library from a CDN (Content Delivery Network).
If this command line switch is enabled, mathjax is copied
from the module mwiki to the static websiste output directory.
The drawback of this approach is the larger disk space and
bandwidth usage.