Disable default equation enumeration within math code blocks.
Disable default equation enumeration inside math code blocks such as
{solution}, {proof} and so on. This behavior automatically adds \notag
to the LaTeX code blocks whithin those blocks. For instance,

````latex
```{solution}
The torque equation of motions is

$$
  \tau_A = \mathcal{I}_A \dot{\omega
$$
````
````

This automatically adds the code "\notag" to the equation code for
disabling equation enumeration in this block. As a result, the
equation code block is turned into

````latex
```{solution}
The torque equation of motions is

$$
 \notag
  \tau_A = \mathcal{I}_A \dot{\omega
$$
````
````

To enable the equation enumeration within the code block
use the directive enumeration

````latex
```{solution}
:enumeration: on

The torque equation of motions is

$$
 \notag
  \tau_A = \mathcal{I}_A \dot{\omega
$$
```
````