Feature - Implement equation derivation foldable section.
Implement equation derivation foldable section with the syntax

````markdown
```{derivation}
According to Euler's law of motion for torque, the angular momentum
and torque are related as

$$
\boldsymbol{\tau} = \frac{d}{dt} \mathbf{L}
$$

Where $\mathbf{L}$ is the angular momentum.
```
````

The label of this foldable block similar to {solution} or {proof} blocks
is rendered to "Equation Derivation" for the English locale. New localizations
for this label can be added by adding new translation in main.js file to the I18N tag
"foldable-math-derivation-block-label".