Feature - Implement custom markdown/JSON syntax for flashcard deck.
Implement the following syntax for flashcard deck. For instance,
the first flashcard of this deck is the entry `["der Osten", "east"]`.
The front side of this flashcard is the german word "der Osten"
and the back side is the corresponding English word "east".

````markdown
```{flashcard}
{
     "title":  "Vocabulary German/English"
   , "entries": [
           [ "der Osten", "east"]
        ,  ["der Tag",    "day"]
        ,  ["der Wochentag", "weekday or day of week"]
        ,  ["der Diestang",  "Tuesday"]
        ,  ["der Mittwoch", "Wednesday"]
       ,   ["die Sicht", "View"]
       ,   ["die Ausgabe", "Issue, example magazine or journal issue"]
       ,   ["die Vorherige Ausgabe", "Previous Issue"]
       ,  ["seher gut", "very good or very well"]
   ]
}
```
````

This custom markdown code block is rendered to a flashcard deck,
where each flashcard is an element of the JSON array entries.
The front side of each flashcard is the first element of each flashcard
in the array entries and the back side (answer) is the corresponding
second element of the corresponding flashcard array entries.
The button view displays all flashcards and their corresponding answers.
The button previous displays the next flashcard in the deck. The prvious button
switches back to the previous flashcard. The checkbox random enables
random mode in which the next flashcard is displayed by the button next
is a random one. This feature is useful for practicing foreign languages or
memorize information.