Feature - Command $ mwiki export now can compile LaTeX on server side using KaTeX
Now the command $ mwiki export with the --compile-latex can build static websistes or self-contained html files with LaTeX formulas embedded as HTML fragments compiled by  KaTeX CLI - Command Line Interface. This feature requires a NodeJS installation and the parent directory of the node executable to be listed in the $PATH environment variable. If nodeJS is not avaiable in or accessible in the command line,
it is possible to set the environment variable MWIKI_NODE_PATH before calling
the command $ mwiki export, example

```sh
$ export MWIKI_NODE_PATH=/path/to/node/js/executable/node
$ mwiki export --wikithpath=./path/to/repo --output=./dist --compile-latex ....
```