Feature - Implement  popup window for displaying LaTeX equation crosss-reference links.
In LaTex, cross-reference hyperliniks to a equation can be defined using the command
 \eqref{unique-label}. Such link refers to a equation of label "unique-label", that can be
set using the command \label in a LaTex code, for instance,

```latex
\label{unique-label}
ax^2 + bx + c = 0
```

When an user clicks at a cross-reference link, the web browser  goes to
that link. However, the user may lose context and jumping to too many
hyperlinks can be cumbersome and time consuming. In order to solve
this issue, this commits implements a popup window that displays
the latex formula refered by such links when the mouse hoovers over it.
This feature makes it easier to read documents with lots forrmulas,
equations derivations and cross references. The window is closed
if the user clicks outside of it.