Feature - implement frontmatter directive latex_renderer for overriding the default LaTeX renderer.
The frontmatter directive latex_renderer allows overriding the default LaTeX renderer
that is set in the settings form page /settings. For instance, if the markdown file of
any MWiki page has the directive

````markdown
```
---
latex_renderer:  mathjax
---

## Quaternions
  .... ... ... ...
## Equations of motion
```
````

This page will be rendered using Mathjax regardless of the default (global) LaTeX
renderer defined in the settings form. The possible choices for this compile-time
directive are "mathjax" for rendering the document with MathJax JavaScript library
or KaTeX for rendering the document using KaTeX javascript library.