Feature - Implement LaTeX input window for the code editor.
Implement LaTeX input window (popup window) in which an user
can type a LaTeX formula or code and get immediate feedback about
how the formula looks like as the it is rendered immediately as the user
types. This feature improves the usability, user experience (UX) by
making it easier and faster to enter LaTeX formulas.