Automatically renew CSRF token for the wiki code editor.
Background: Due to the CSRF (Cross-Site Request Forgery)
token protection, the editor was not able to send data
through the REST API endpoints if the editor session was
long enough and caused CSRF token expiration. The solution
adopted was to automatically renew the CSRF token every
30 minutes by making GET http requests (Ajax calls) to the
API endpoint /api/token, which issues a new CSRF token for
logged in users.