Bugfix - Fix I8N internationalization of elements with string interpolation.
The internationalization of DOM elements with I18N tag with
string interpolation did not work due to the lack of idempotency
of the function setLocaleI18N. A function (procedure, subroutine)
is said to be idempotent if multiple runs of this function yields the
same effect.

Bug Description: For instance, a DOM element such as

```html
<h1 data-i18n="page-title">[i18n] Web Hosting</h1>
```

where the corresponding value of this i18n tag  "Editing" was
being localized to English as

```html
<h1 data-i18n="page-title">Editing</h1>
```

Instead of "Editing Web Hosting". This glitch happened because the setLocaleI18n()
was not idempotent. The first run of this setLocaleI18() would correctly localize the DOM
element to "Editing Web Hosting". A second run of this function would
replace the entire textContent attribute by the value of the i18n tag "Editing".

Solution: Store the i18n text template in the dataset attribute

```
domElement.dataset.textContentI18N = "[i18n] Web Hosting"
```

and always use the text of this custom DOM attribute for replacing
textContent attribute of the DOM element.