Feature - Printer icon displays page in desktop mode in mobile devices for better printing.
The printer icon on the topbar (page header) opens the current MWiki
page in  a new tab, displays the it in desktop mode and calls the
function window.printer() in mobile devices. This feature allows printing or
exporting to PDF from mobile devices in the same way as printing from
desktop devices.